+++ api/blueprints/user/views/__init__.py
@@ -1,0 +1,1 @@
+from . import user  # noqa F401
